Moved files to the right places
